+++ core/php/AbeilleCmd-Tuya.php
@@ -125,45 +125,71 @@
         return $tId;
     }
 
-    // Use cases: E004+ED00 clusters support (Moes universal remote)
+    require_once __DIR__.'/../class/AbeilleCmdProcess.class.php';
+
+    // Use cases: ED00 cluster support (Moes universal remote)
     function tuyaZosung($net, $addr, $ep, $cmd, $data) {
-        cmdLog('debug', '  tuyaZosung() cmd');
+        cmdLog('debug', "  tuyaZosung(net=${net}, addr=${addr}, ep=${ep}, cmd=${cmd})");
 
-        $irMsg = array(
-            'key_num' => 1,
-            'delay' => 300,
-            'key1' => array (
-                'num' => 1,
-                'freq' => 38000,
-                'type' => 1,
-                'key_code' => $data,
-            ),
-        );
-        $irMsgJson = json_encode($irMsg);
+        if ($cmd == '00') {
+            $irMsg = array(
+                'key_num' => 1,
+                'delay' => 300,
+                'key1' => array (
+                    'num' => 1,
+                    'freq' => 38000,
+                    'type' => 1,
+                    'key_code' => $data,
+                ),
+            );
+            $irMsgJson = json_encode($irMsg);
 
-        // Cmd 00 reminder
-        // {name: 'seq', type: DataType.uint16},
-        // {name: 'length', type: DataType.uint32},
-        // {name: 'unk1', type: DataType.uint32},
-        // {name: 'unk2', type: DataType.uint16},
-        // {name: 'unk3', type: DataType.uint8},
-        // {name: 'cmd', type: DataType.uint8},
-        // {name: 'unk4', type: DataType.uint16},
-        $seq = '0012'; // TODO
-        $length = sprintf("%04X", strlen($irMsgJson));
-        $unk1 = '00000000';
-        $unk2 = 'e004';
-        $unk3 = '01';
-        $cmd = '02';
-        $unk4 = '0000';
+            // Saving message to send
+            $GLOBALS['zosung_msg'] = array(
+                'message' => $data
+            );
 
-        $seq = AbeilleTools::reverseHex($seq);
-        $length = AbeilleTools::reverseHex($length);
-        $unk1 = AbeilleTools::reverseHex($unk1);
-        $unk2 = AbeilleTools::reverseHex($unk2);
-        $unk4 = AbeilleTools::reverseHex($unk4);
-        $data = '00'.$seq.$len.$unk1.$unk2.$unk3.$cmd.$unk4;
-        // msgToCmd(PRIO_NORM, "Cmd".$net."/".$addr."/cmd-Generic", "ep=".$ep."&clustId=ED00&cmd=00&data=${data}");
+            // Cmd ED00-00 reminder
+            // {name: 'seq', type: DataType.uint16},
+            // {name: 'length', type: DataType.uint32},
+            // {name: 'unk1', type: DataType.uint32},
+            // {name: 'unk2', type: DataType.uint16},
+            // {name: 'unk3', type: DataType.uint8},
+            // {name: 'cmd', type: DataType.uint8},
+            // {name: 'unk4', type: DataType.uint16},
+            $seq = '0012'; // TODO
+            $length = sprintf("%08X", strlen($irMsgJson));
+            $unk1 = '00000000';
+            $unk2 = 'e004';
+            $unk3 = '01';
+            $cmd = '02';
+            $unk4 = '0000';
 
+            $seq = AbeilleTools::reverseHex($seq);
+            $len = AbeilleTools::reverseHex($length);
+            $unk1 = AbeilleTools::reverseHex($unk1);
+            $unk2 = AbeilleTools::reverseHex($unk2);
+            $unk4 = AbeilleTools::reverseHex($unk4);
+            $data = $seq.$len.$unk1.$unk2.$unk3.$cmd.$unk4;
+
+            $header = array(
+                'net' => $net,
+                'addr' => $addr,
+                'ep' => $ep,
+                'clustId' => 'ED00',
+                'clustSpecific' => true, // Not a general frame
+                // 'manufCode' => 'xxxx'
+                'cmd' => '00'
+            );
+            AbeilleCmdProcess::sendRawMessage($header, $data);
+        } else if ($cmd == '03') {
+            // Cmd 03 reminder
+            // {name: 'zero', type: DataType.uint8},
+            // {name: 'seq', type: DataType.uint16},
+            // {name: 'position', type: DataType.uint32},
+            // {name: 'msgpart', type: DataType.octetStr},
+            // {name: 'msgpartcrc', type: DataType.uint8},
+            // $data = '00'.$seq.$position.$msgpart.$msgpartcrc;
+        }
     }
 ?>